% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_defaults.R
\name{assign_defaults}
\alias{assign_defaults}
\alias{assign_defaults_addin}
\alias{assign_defaults}
\title{Set function arguments to defaults}
\usage{
assign_defaults_addin()

assign_defaults(fun)
}
\arguments{
\item{fun}{A function object or a character string naming the function. See
\code{\link{formals}} for details.}
}
\description{
Set arguments for a function to their default values in the global
environment. The addin takes \code{fun} from current text selection.
}
\details{
The goal is to make it easier to walk through the function body as top-level
code for development and debugging. Arguments with no defaults and \code{...}
are silently ignored. In order for this to work, the function needs to be
defined through one of these methods:
\itemize{
\item \strong{RStudio: Run the current function definition}: With cursor in
the function, do Ctrl+Alt+F (Windows/Linux) or Command+Option+F (Mac).
\item \strong{Source the file with function definition}: Visit file in
RStudio editor and do Ctrl+Shift+S (Windows/Linux) or Command+Shift+S (Mac).
Or use \code{\link{source}()}.
\item \strong{\code{devtools::load_all()}}: Sensible if function is part of a
package and you're developing with \code{devtools}.
\item \strong{\code{library(...)}}: Sensible if function is an exported
function in an installed package.
}
}
