% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jadd-package.r
\name{assign_defaults_addin}
\alias{assign_defaults_addin}
\title{Set function arguments to defaults}
\usage{
assign_defaults_addin()
}
\description{
Set function arguments in selection to their default values in the global
environment. Objective is to aid function development and debugging. Some
sloppiness in selection in tolerated. Arguments with no defaults are silently
ignored -- it is assumed the developer will set them to something sensible
"by hand". Also \code{...} is dropped.
}

