% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudioapi.R
\name{rstudioapi_addin}
\alias{rstudioapi_addin}
\title{Inspect information available via rstudioapi}
\usage{
rstudioapi_addin()
}
\description{
Prints the output of these calls to the rstudioapi package:
\itemize{
\item getActiveProject()
\item getActiveDocumentContext()
\item getSourceEditorContext()
\item getConsoleEditorContext()
\item primary_selection(getSourceEditorContext()
}
}
